Move CFL check to a separate class
