+++ applications/sintering/include/pf-applications/sintering/operator_advection.h
@@ -114,104 +114,6 @@ namespace Sintering
     }
 
     void
-    precompute_cell_diameters()
-    {
-      cell_diameters.resize(this->matrix_free.n_cell_batches());
-
-      std::pair<unsigned int, unsigned int> range{
-        0, this->matrix_free.n_cell_batches()};
-
-      for (auto cell = range.first; cell < range.second; ++cell)
-        {
-          unsigned int i = 0;
-
-          for (; i < this->matrix_free.n_active_entries_per_cell_batch(cell);
-               ++i)
-            {
-              const auto icell = this->matrix_free.get_cell_iterator(cell, i);
-              cell_diameters[cell][i] = icell->diameter();
-            }
-
-          // Assign max double for irregular cells
-          for (; i < VectorizedArrayType::size(); ++i)
-            cell_diameters[cell][i] = std::numeric_limits<
-              typename VectorizedArrayType::value_type>::max();
-        }
-    }
-
-    bool
-    check_courant(const double dt) const
-    {
-      MyScope scope(this->timer,
-                    "advection_op::check_courant",
-                    this->do_timing);
-
-      AssertThrow(
-        cell_diameters.size() == this->matrix_free.n_cell_batches(),
-        ExcMessage(
-          std::string(
-            "The number of precomputed cell diameters is inconsistent") +
-          std::string(" with the number of current cell batches: ") +
-          std::to_string(cell_diameters.size()) + std::string(" and ") +
-          std::to_string(this->matrix_free.n_cell_batches())));
-
-      FECellIntegrator<dim, 1, Number, VectorizedArrayType> phi(
-        this->matrix_free, this->dof_index);
-
-      VectorizedArrayType zeros(0.0);
-      VectorizedArrayType ones(1.0);
-
-      std::pair<unsigned int, unsigned int> range{
-        0, this->matrix_free.n_cell_batches()};
-
-      AdvectionVelocityData<dim, Number, VectorizedArrayType> advection_data(
-        advection_mechanism, data);
-
-      bool status = true;
-
-      for (auto cell = range.first; cell < range.second; ++cell)
-        {
-          if (!status)
-            break;
-
-          phi.reinit(cell);
-
-          // Reinit advection data for the current cells batch
-          advection_data.reinit(cell);
-
-          const auto grain_to_relevant_grain =
-            this->data.get_grain_to_relevant_grain(cell);
-
-          for (unsigned int ig = 0; ig < n_grains() && status; ++ig)
-            if (advection_data.has_velocity(ig))
-              for (unsigned int q = 0; q < phi.n_q_points; ++q)
-                {
-                  const auto &velocity_ig =
-                    advection_data.get_velocity(ig, phi.quadrature_point(q));
-
-                  // Check Courant condition
-                  const auto cdt = velocity_ig.norm() * dt;
-
-                  const auto courant =
-                    compare_and_apply_mask<SIMDComparison::less_than>(
-                      cdt, cell_diameters[cell], zeros, ones);
-
-                  if (courant.sum() > 0)
-                    {
-                      status = false;
-                      break;
-                    }
-                }
-        }
-
-      status = static_cast<bool>(Utilities::MPI::min(
-        static_cast<unsigned int>(status),
-        this->matrix_free.get_dof_handler().get_mpi_communicator()));
-
-      return status;
-    }
-
-    void
     do_update() override
     {
       if (this->matrix_based)
@@ -491,28 +393,5 @@ namespace Sintering
     const GrainTracker::Tracker<dim, Number>              &grain_tracker;
 
     AdvectionMechanism<dim, Number, VectorizedArrayType> &advection_mechanism;
-
-    AlignedVector<VectorizedArrayType> cell_diameters;
-  };
-
-  class ExcCourantConditionViolated : public dealii::ExceptionBase
-  {
-  public:
-    ExcCourantConditionViolated() = default;
-
-    virtual ~ExcCourantConditionViolated() noexcept override = default;
-
-    virtual void
-    print_info(std::ostream &out) const override
-    {
-      out << message() << std::endl;
-    }
-
-    std::string
-    message() const
-    {
-      return "The Courant condition was violated. "
-             "The advection velocity is too high.";
-    }
   };
 } // namespace Sintering
